content/base/test/test_bug431833.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=431833
-->
<head>
  <title>Test for Bug 431833</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=431833">Mozilla Bug 431833</a>
<p id="display">
 <iframe id="f1" src="data:text/html,1"></iframe>
 <iframe id="f2" src="data:text/html,2"></iframe>
 <iframe id="f3" src="data:text/html,<iframe id='f4' src='data:text/html,3'></iframe>"></iframe>
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 431833 **/

var str = "FAIL";
function test(e) {
  if (str == "FAIL")
    str = "";

  str += "PASS: " + e.target.id + '. '; 
}

window.addEventListener('DOMFrameContentLoaded',test,true);

SimpleTest.waitForExplicitFinish();

addLoadEvent(function() {
  is(str, "PASS: f1. PASS: f2. PASS: f3. PASS: f4. ",
     "DOMFrameContentLoaded events didn't fire?");
  });

addLoadEvent(SimpleTest.finish);
</script>
</pre>
</body>
</html>
